Add files via upload
This project focuses on the "Seoul Bike Sharing Demand Data Set" (Kang et al., 2018)
which provides valuable insights into bike rentals in Seoul, South Korea. The goal of
the project is to explore the dataset, develop a predictive model, and gain a better
understanding of the factors influencing bike rental counts. By examining variables such
as temperature, humidity, wind speed, and seasonal variations, we aim to contribute to
the improvement of bike sharing systems and resource allocation.
To solve the problem at hand, we will follow a systematic approach. First, we will
perform exploratory data analysis to comprehensively understand the dataset. This step
will involve assessing the characteristics and distribution of the variables, identifying any
missing values or outliers, and obtaining initial insights into their relationships.
Next, we will conduct descriptive analysis using techniques such as scatter plots or
correlation plots. This analysis will provide visualizations and statistical measures to
explore the relationships between variables. By identifying potential trends or patterns,
we can gain insights into the factors that impact bike rental counts.
Succeeding the descriptive analysis, we will develop a regression model to predict the
number of bike rentals. First, we will use a linear regression model using all available independent
variables. Through this model, we can assess the significance of each variable
and evaluate the model’s performance.
To enhance the interpretability and efficiency of the model, we will use the best subset
selection technique that will guide us in identifying a subset of explanatory variables that
significantly influence bike rental counts. We will consider criteria such as AIC, BIC,
adjusted-R2, or Mallows’ Cp values to aid in the selection process. Once the selected
regression model is established, we will evaluate its performance. This evaluation will
involve analyzing residual plots to assess linearity, heteroskedasticity, and normality.
Additionally, we will examine the variance inflation factor (VIF) to detect and address
any multicollinearity issues that may arise. The results of the regression model will be
interpreted to understand the impact of the selected variables on the bike rental counts.
In conclusion, this project aims to explore the bike rental data set and develop a regression
model to predict a bike rental counts. By analyzing the relationships between
variables, selecting significant predictors, and interpreting the results, we seek to contribute
to the improvement of bike-sharing systems in Seoul.